All Couche finish (no test)

+++ TPEnchere/src/main/java/org/eni/encheres/utilitaire/FicheMethodeBool.java
@@ -1,15 +1,15 @@
 package org.eni.encheres.utilitaire;
 
 public class FicheMethodeBool {
-	public static boolean bitToBool(int bit) {
+	public static boolean bitToBool(byte bit) {
 		boolean vretour = false;
 		if(bit == 1) {
 			vretour = true;
 		}
 		return vretour;
 	}
-	public static int boolToBit(boolean bool) {
-		int vretour = 0;
+	public static byte boolToBit(boolean bool) {
+		byte vretour = 0;
 		if(bool) {
 			vretour = 1;
 		}
